add "editor" module
