+++ src/neuroglancer/viewer_state.ts
@@ -15,7 +15,7 @@
  */
 
 import {LayerManager, MouseSelectionState} from 'neuroglancer/layer';
-import {EditorState} from 'neuroglancer/viewer_editors';
+import {EditorState} from 'neuroglancer/editor/state';
 import {TrackableValue} from 'neuroglancer/trackable_value';
 import {NavigationState} from 'neuroglancer/navigation_state';
 import {TrackableBoolean} from 'neuroglancer/trackable_boolean';
